+++ packages/reactabular-column-extensions/src/extensions/highlight-cell.js
@@ -1,0 +1,20 @@
+import * as search from 'searchtabular';
+
+function highlightCell() {
+  return {
+    match({ cell }) {
+      return cell.highlight;
+    },
+    evaluate() {
+      return {
+        cell: {
+          formatters: [
+            search.highlightCell
+          ]
+        }
+      };
+    }
+  };
+}
+
+export default highlightCell;