+++ packages/reactabular-column-extensions/src/index.js
@@ -1,0 +1,6 @@
+export { default as bind } from './bind';
+export { default as draggableHeader } from './extensions/draggable-header';
+export { default as highlightCell } from './extensions/highlight-cell';
+export { default as resizableHeader } from './extensions/resizable-header';
+export { default as sortableHeader } from './extensions/sortable-header';
+export { default as toggleChildrenCell } from './extensions/toggle-children-cell';